+++ src/grammar/GrammarBackedCell.ts
@@ -58,6 +58,9 @@ abstract class AbstractGrammarBackedCell<T> {
     const definition = this._getCellTypeDefinition()
     let words = definition ? definition.getAutocompleteWordOptions(this._program) : []
 
+    const runTimeOptions = this._node.getRunTimeEnumOptions(this)
+    if (runTimeOptions) words = runTimeOptions.concat(words)
+
     if (partialWord) words = words.filter(word => word.includes(partialWord))
     return words.map(word => {
       return {
@@ -82,6 +85,8 @@ abstract class AbstractGrammarBackedCell<T> {
   protected abstract _isValid(): boolean
 
   isValid(): boolean {
+    const runTimeOptions = this._node.getRunTimeEnumOptions(this)
+    if (runTimeOptions) return runTimeOptions.includes(this._word)
     return this._getCellTypeDefinition().isValid(this._word, this._node.getProgram()) && this._isValid()
   }